geometry-mesh-fluent/requirements_24.1.txt
ansys-geometry-core[all]==0.5.3
ansys-meshing-prime==0.5.1
